requirements.txt
# -----------------------------
# Core dependencies
# -----------------------------
fastapi
uvicorn
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic
httpx

# -----------------------------
# LangChain & AI
# -----------------------------
langchain
langchain[google-genai]
langchain-community
langchain-chroma
langgraph
chromadb
sentence-transformers
langchain_tavily
# -----------------------------
# Task scheduling
# -----------------------------
APScheduler

# -----------------------------
# Telegram bot
# -----------------------------
python-telegram-bot>=20.4
python-telegram-bot[webhooks]

# -----------------------------
# Visualization / Statistics
# -----------------------------
matplotlib

# -----------------------------
# Development / utilities
# -----------------------------

pytrends